gpgme/ 2001-11-02  Marcus Brinkmann  <marcus@g10code.de>
	* rungpg.c (pipemode_copy): Change type of NBYTES to size_t.

	* key.c: Include string.h.
	* data.c: Likewise.
	* recipient.c: Likewise.

tests/
2001-11-02  Marcus Brinkmann  <marcus@g10code.de>

	* t-data.c: Include stddef.h.
	(read_once_test): Change type of READ to size_t.
	(read_test): Likewise.
	(write_test): Likewise.

+++ gpgme/rungpg.c
@@ -1275,7 +1275,7 @@ static GpgmeError
 pipemode_copy (char *buffer, size_t length, size_t *nread, GpgmeData data )
 {
     GpgmeError err;
-    int nbytes;
+    size_t nbytes;
     char tmp[1000], *s, *d;
 
     /* we can optimize this whole thing but for now we just
